requirements.txt
absl-py==0.9.0
aenum==2.2.3
aiohttp==3.6.2
alembic==1.4.2
asgiref==3.2.10
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cliff==3.3.0
cmaes==0.5.1
cmd2==1.2.1
colorama==0.4.3
colorlog==4.1.0
comet-git-pure==0.19.16
comet-ml==3.1.12
configobj==5.0.6
cycler==0.10.0
dill==0.3.2
Django==3.0.8
dnspython==2.0.0
eventlet==0.25.2
everett==1.0.2
firebase-admin==4.3.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
future==0.18.2
futures==3.1.1
gast==0.3.3
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.30.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
livereload==2.6.2
lunr==0.5.8
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mkdocs==1.1.2
monotonic==1.5
msgpack==1.0.0
multidict==4.7.6
nest-asyncio==1.4.0
netifaces==0.10.9
nltk==3.5
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
optuna==1.5.0
pandas==1.0.5
pbr==5.4.5
pdoc3==0.8.4
phonenumbers==8.12.6
prettytable==0.7.2
prompt-toolkit==1.0.14
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyInquirer==1.0.3
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
stevedore==3.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.7.0
tensorflowjs==2.0.1.post1
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.47.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer==2.0.1
yarl==1.4.2
zipp==3.1.0
